reading/search_indexes.py
from haystack import indexes

from reading.models import Page


class BookPageIndex(indexes.SearchIndex, indexes.Indexable):
    text = indexes.CharField(document=True, use_template=True)

    def get_model(self):
        return Page
