requirements.txt
anyjson==0.3.3
asn1crypto==0.24.0
boto3==1.9.143
botocore==1.12.143
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
construct==2.5.3
cryptography==2.6.1
defusedxml==0.6.0
Django==2.2.1
django-allauth==0.39.1
django-environ==0.4.5
django-haystack==2.8.1
django-storages==1.7.1
django-tinymce==2.8.0
docutils==0.14
future==0.17.1
gunicorn==19.9.0
idna==2.8
jmespath==0.9.4
oauthlib==3.0.1
paypalrestsdk==1.13.1
pefile==2019.4.18
Pillow==6.0.0
psycopg2-binary==2.8.2
pycparser==2.19
pyOpenSSL==19.0.0
PyPDF2==1.26.0
python-dateutil==2.8.0
python-ptrace==0.9.3
python3-openid==3.1.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
s3transfer==0.2.0
six==1.12.0
sqlparse==0.3.0
stripe==2.27.0
urllib3==1.24.3
Whoosh==2.7.4
celery==4.3.0
Wand==0.5.3
django-cities-light==3.5.0
pyPdf==1.13
pytesseract==0.2.6
pyocr==0.7
django-el-pagination==3.2.4
